issue-2631: move some code from main.cpp of filestore-client
